Move SimplePianoRollMultiChannel from view/util to view to view
